requirements.txt
aerich==0.7.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
atlastk==0.13.3
attrs==24.2.0
bcrypt==4.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dictdiffer==0.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.0
email_validator==2.1.2
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-users==13.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
makefun==1.15.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
orjson==3.10.10
passlib==1.7.4
postgres==4.0
propcache==0.2.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pwdlib==0.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sql==2022.4.0
starlette==0.41.2
tomlkit==0.13.2
tortoise==0.1.1
tortoise-orm==0.21.7
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.0